Refactor Implementation of HttpFake and MetaFake out of header file
