+++ src/aktualizr_primary/primary_secondary_registration_test.cc
@@ -2,6 +2,7 @@
 
 #include "httpfake.h"
 #include "libaktualizr/aktualizr.h"
+#include "metafake.h"
 #include "secondary.h"
 #include "uptane_test_common.h"
 #include "utilities/utils.h"
@@ -127,7 +128,7 @@ int main(int argc, char** argv) {
 
   TemporaryDirectory tmp_dir;
   fake_meta_dir = tmp_dir.Path();
-  MetaFake meta_fake(fake_meta_dir);
+  CreateFakeRepoMetaData(fake_meta_dir);
 
   return RUN_ALL_TESTS();
 }
